+++ src/types.ts
@@ -1,0 +1,15 @@
+export interface ANFSettings {
+	enableImage: boolean;
+	image: string;
+	enableAudio: boolean;
+	audio: string;
+	enableVideo: boolean;
+	video: string;
+	enablePdf: boolean;
+	pdf: string;
+	connector: string;
+	exportCurrentDeletion: boolean;
+	exportUnusedDeletion: boolean;
+	copyPath: boolean;
+	copyPathMode: string;
+}